Created remote Client for Github repos Built tests for the remote Client Created parser for the Link header response Updated Github Repos Coordinator

+++ App/Sources/Globals/Models/GithubReposPage.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct GithubReposPage: Decodable {
-    let totalCount: UInt16
+    let totalCount: UInt
     let incompleteResults: Bool
     let items: [GithubRepoModel]
 }
